webperimental: let the Leader heal its units.
Give the Leader unit the ability to do "Heal Unit" to domestic and allied
light weight units, including it self. The ability to heal it self and any
light weight escorts makes it a bit harder to kill the Leader. It could also
give a (small) incentive to take the Leader into the field in the beginning
of the game (when most land units have the LightWeight unit type flag).

See patch #7775

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@34020 a0f10bec-cc02-0410-94fc-a9cfff90b4cd