ai attacker want: check occupation action...
when deciding how much a unit of a certain type is wanted (how much utility
it has). A unit that can conquer a city is desired 10x the target city worth
if the invasion has other units that can attack but no units that can
conquer the city.

See patch #7778

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@34038 a0f10bec-cc02-0410-94fc-a9cfff90b4cd