Auto attack is performed by the rules.
When the autoattack setting is enabled a server side auto attack happens
automatically. This means that auto attack isn't player performed. The
player doesn't control if his units (or a particular unit) should auto
attack or not. This means that auto attack isn't controlled by a server side
agent. The autoattack setting is a rule. Declare that the actions performed
by it is done by the rules.

See patch #7811

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@34108 a0f10bec-cc02-0410-94fc-a9cfff90b4cd