Introduce the new utype flag "Provoking".
A unit considering if it should autoattack a unit that just moved next to it
will normally decide not to attack if it has better odds when defending
against it than when attacking it. Exceptions are made for diplomats and for
transports. Move this exception to the new unit type flag "Provoking".

See patch #7943

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@34426 a0f10bec-cc02-0410-94fc-a9cfff90b4cd