An unknown target can't make an action possible.
An action can be known to be impossible because of the actor unit. When this
is true it doesn't matter what the target is.

Example: It doesn't matter if the unseen distant tile the unit wants to
bombard has units or not when it is out of moves and the rules requires a
bombarding unit to have a move fragment left.

Have the action probability calculation for unit stack targeted actions
return impossible rather than unknown when the actor unit is known to make
the action impossible.

See bug #25281

git-svn-id: svn://svn.gna.org/svn/freeciv/trunk@34497 a0f10bec-cc02-0410-94fc-a9cfff90b4cd